backend/requirements.txt
Django>=5.1.2		# Django web framework
djangorestframework>=3.14.0		# Django REST framework
django-environ>=0.11.2		# Environment variables for Django
psycopg2>=2.9.10		# PostgreSQL database adapter for postgres engine
djangorestframework-simplejwt>=5.3.0		# JWT authentication for Django REST framework
django-cors-headers>=4.3.0		# CORS headers for Django
redis>=5.2.0		# Redis client for Python
django-redis>=5.4.0		# Redis cache backend for Django
